Split test cases per provider
