+++ tests/test_oauth2_implicit_flow_azure_active_directory.py
